requirements.txt
gitpython==3.1.43
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
openai>=1.0.0
streamlit==1.41.1
python-dotenv==1.0.1
